--- pages/agora/auth.tsx
@@ -1,10 +1,0 @@
-import { NextPage } from "next";
-import { Index } from "../../components/pages/agora/AgoraAuthPage";
-
-interface Props {}
-
-const AuthPageContainer: NextPage<Props> = () => {
-  return <Index />;
-};
-
-export default AuthPageContainer;
