+++ styles/profile/AuthModal.module.scss
@@ -1,15 +1,7 @@
 @import "styles/Buttons";
 
-.container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
 .form {
-  padding: 45px 270px;
+  padding: 46px 126px;
   border: var(--color-main) 1px solid;
 
   .input {
@@ -44,6 +36,7 @@
     line-height: 1;
     padding: 5px 0;
     margin: 5px 0;
+    font-size: 15px;
   }
 
   .label {
@@ -55,6 +48,7 @@
     border-color: #c4c4c4 !important;
     color: #c4c4c4;
     height: 30px;
+    width: 69px;
 
     &:not(:first-of-type) {
       margin-left: 4px;
